HBASE-27279 Make SslHandler work with SaslWrapHandler/SaslUnwrapHandler (#4705)
(cherry picked from commit 2b9d36869ffe765256d4a7b6a74b95a8200678df)

Conflicts:
	hbase-server/src/test/java/org/apache/hadoop/hbase/security/TestNettyTlsIPCRejectPlainText.java
	hbase-server/src/test/java/org/apache/hadoop/hbase/security/TestSecureIPC.java
	hbase-server/src/test/java/org/apache/hadoop/hbase/security/TestTlsWithKerberos.java
